% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrade.R
\name{calcGrade}
\alias{calcGrade}
\title{Calculates overall course grade}
\usage{
calcGrade(courseDF, courseCategoryDF)
}
\arguments{
\item{courseDF}{Course data frame generated by uploadAssignments().}

\item{courseCategoryDF}{Category data frame generated by createCourseWeights().}
}
\value{
Overall course grade.
}
\description{
Calculates overall course grade
}
\examples{
calcGrade(PSY2012, PSY2012Categories)
}
