% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFinalGradeNeeded.R
\name{calculateFinalGradeNeeded}
\alias{calculateFinalGradeNeeded}
\title{Calculate grade needed on a final for a course}
\usage{
calculateFinalGradeNeeded(currentGrade, goalGrade, finalWeight)
}
\arguments{
\item{currentGrade}{Current course grade.}

\item{goalGrade}{Grade you would like to achieve.}

\item{finalWeight}{Weight of the final.}
}
\value{
Grade needed on final to achieve desired course grade.
}
\description{
Calculate grade needed on a final for a course
}
\examples{
calculateFinalGradeNeeded(87, 92, 0.2)
}
