% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAssignments.R
\name{createAssignments}
\alias{createAssignments}
\title{Creates course assignments using Shiny form}
\usage{
createAssignments()
}
\value{
Form responses.
}
\description{
Input all course assignments for all categories at once. Then use uploadAssignments().
}
