% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCourseWeights.R
\name{createCourseWeights}
\alias{createCourseWeights}
\title{Creates course categories data frame}
\usage{
createCourseWeights(courseName, categoryNames, categoryWeights)
}
\arguments{
\item{courseName}{Name of course. Do not include spaces or anything that would not be allowed in a variable name}

\item{categoryNames}{Character vector containing category names for given course.}

\item{categoryWeights}{Character vector containing respective weights for each category for given course. Format as decimal.}
}
\value{
Course Categories data frame.
}
\description{
Creates course categories data frame
}
\examples{
createCourseWeights("PSY2012", c("Papers", "Quizzes", "Midterm", "Final"), c(0.2,0.4,0.2,0.2))
}
