% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadAssignments.R
\name{uploadAssignments}
\alias{uploadAssignments}
\title{Uploads assignments to course data frame}
\usage{
uploadAssignments(courseName)
}
\arguments{
\item{courseName}{Name of course.}
}
\value{
Course data frame filled with inputted assignments.
}
\description{
Must be used after createAssignments().
}
